requirements.txt
# Web Framework
Flask==2.3.3
Flask-Cors==4.0.0
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
alembic==1.12.0
Flask-Migrate==4.0.5

# Authentication
Flask-JWT-Extended==4.5.3
passlib==1.7.4
bcrypt==4.0.1

# Caching and task queue
redis==5.0.0
celery==5.3.4

# API and Serialization
flask-restx==1.1.0
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Utils
python-dotenv==1.0.0
gunicorn==21.2.0
pillow==10.0.1
pydantic==2.3.0
email-validator==2.0.0